REFACTOR: move CustomerPortal overrides in a new my2 file - created my2_portal.py - moved _update_password
